refactor: extract settings modules
